// src/types/config.ts
export type ProfileConfig = {
	avatar?: string;
	name: string;
	fancyName: string;
	bio?: string;
	links: {
		name: string;
		url: string;
		icon: string;
		showName?: boolean;
	}[];
};